+++ src/IfcJsonHelper.cs
@@ -11,54 +11,52 @@ namespace Bingosoft.Net.IfcMetadata
         internal static void ToJson(FileInfo jsonTargetFile, ref MetadataExtractor metadata)
         {
             using var stream = File.OpenWrite(jsonTargetFile.FullName);
-            using (var writer = new Utf8JsonWriter(stream, Jwo))
+            using var writer = new Utf8JsonWriter(stream, Jwo);
+            writer.WriteStartObject();
+
+            writer.WriteString("id", metadata.Id);
+            writer.WriteString("projectId", metadata.ProjectId);
+            writer.WriteString("author", metadata.Author);
+            writer.WriteString("createdAt", metadata.CreatedAt);
+            writer.WriteString("schema", metadata.Schema);
+            writer.WriteString("creatingApplication", metadata.CreatingApplication);
+
+            writer.WriteStartObject("metaObjects");
+
+            foreach (var item in metadata.MetaObjects)
             {
-                writer.WriteStartObject();
+                writer.WriteStartObject(item.Id);
 
-                writer.WriteString("id", metadata.Id);
-                writer.WriteString("projectId", metadata.ProjectId);
-                writer.WriteString("author", metadata.Author);
-                writer.WriteString("createdAt", metadata.CreatedAt);
-                writer.WriteString("schema", metadata.Schema);
-                writer.WriteString("creatingApplication", metadata.CreatingApplication);
+                writer.WriteString("id", item.Id);
+                writer.WriteString("name", item.Name);
+                writer.WriteString("type", item.Type);
+                writer.WriteString("parent", item.Parent);
 
-                writer.WriteStartObject("metaObjects");
-
-                foreach (var item in metadata.MetaObjects)
+                if (item.PropertyIds?.Length > 0)
                 {
-                    writer.WriteStartObject(item.Id);
+                    writer.WriteStartArray("properties");
 
-                    writer.WriteString("id", item.Id);
-                    writer.WriteString("name", item.Name);
-                    writer.WriteString("type", item.Type);
-                    writer.WriteString("parent", item.Parent);
-
-                    if (item.PropertyIds?.Length > 0)
+                    foreach (var propertyId in item.PropertyIds)
                     {
-                        writer.WriteStartArray("properties");
-
-                        foreach (var propertyId in item.PropertyIds)
-                        {
-                            writer.WriteStringValue(propertyId);
-                        }
-
-                        writer.WriteEndArray();
-                    }
-                    else
-                    {
-                        writer.WriteNull("properties");
+                        writer.WriteStringValue(propertyId);
                     }
 
-                    writer.WriteString("material_id", item.Material);
-                    writer.WriteString("type_id", item.TypeId);
-
-                    writer.WriteEndObject();
+                    writer.WriteEndArray();
+                }
+                else
+                {
+                    writer.WriteNull("properties");
                 }
 
-                writer.WriteEndObject();
+                writer.WriteString("material_id", item.Material);
+                writer.WriteString("type_id", item.TypeId);
 
                 writer.WriteEndObject();
             }
+
+            writer.WriteEndObject();
+
+            writer.WriteEndObject();
         }
     }
 }